integration-cleanup: use repository file structure on web server
the web folder on the server is now our repository how it is in git. So
you can test on your own machine as on the server and we do not need to
copy files all the time between the repo and the web folder.
Furthermore, we can just do git pull and there is the newest version on
the webserver.
We only exclude the models from the repository, because they would blow
it up.