integration-cleanup: model-data for testing purpose as on server
I changed the file-structure of the models folder to match the structure on
the server. In the repo we only want to have a small set of small test
models. The big models and all the models someone uploaded sometime are
only on the server to keep the git repository small and fast.

I also removed the huge commit with the helmet models from the history
to fix the problems we had with git via http.