Viewer: update on viewpoint change to work on touch/reset/...
Changed the used event to send the pos back to ViewpointChanged, because
there are many more cases than mousedown mouseup (e.g. using a touch
device, resetting the view with 'R','A' or the button in the toolbar).
Therefore, I also disabled sending as long as the interpolation is
working.