Registration: An email is sent to a fix email adress, if someone creates an account.
The email contains a link to a confirmation page. The confirmation page displays infos.
TODO: The confirmation flag in the DB has to be set.

src/config/ now contains a config.php, where the path to the root directory of the web service has to be described. This is used to provide the correct link in the mail.